refactor: 마이페이지 suspense 구조 변경
